Make WSGI script path configurable at runtime
Initializtion of formula is moved to a position more ahead so that it can be merged into the arguments list. Two arguments are now available to override the formula's default WSGI path at runtime.

Also changed the location of gunicorn.pid so that it is out of the project directory (but still in the virtualenv). The Gunicorn startup command is trimmed a bit, too.